+++ src/winch_node.py
@@ -2,6 +2,7 @@
 import rospy
 from sensor_msgs.msg import Joy
 from std_msgs.msg import Int32
+from std_msgs.msg import Float32
 from std_msgs.msg import Bool
 from boat_nav.msg import BoatState
       
@@ -80,7 +81,7 @@ def listener():
     rospy.init_node('joy_to_winch', anonymous=True)
     rospy.Subscriber('joy', Joy, joy_callback)
     rospy.Subscriber('boat_state', BoatState, state_callback)
-    rospy.Subscriber('anemometer', Int32, anemometer_callback)
+    rospy.Subscriber('anemometer', Float32, anemometer_callback)
     rospy.spin()
 
 if __name__ == '__main__':
